Fix layout on small (<1100px width) screens
Closes #25

When the window width was <1100px the left navigation list was cut off.

This was happening due to `marginLeft: -${spaceMedium}` set on rows.
My best guess is that this shifting to the left was balancing out the github ribbon.

I removed these margins and instead padded the wrapping container from the right.
I also padded it from the left, so that there is some space between the screen
edge and the list.

On normal screens the result looks almost (+- few pixels) identical to what it
used to be, and on smaller screens it's now much much better.