race operation: changed relative timings
The previous example was a bit ambigous: 
It could have said: The first responding observable is passed on.
Or it could  have said: the first incoming response is taken as first in the result. Then the second element in the result is taken from whatever observable provides its second element before all others provide their second element.